hurb_trial/requirements.txt
tortoise-orm==0.17.7
fastapi==0.68.1
uvicorn==0.15.0
asyncpg==0.24.0
black==21.8b0
flake8==3.9.2
requests==2.26.0
